Remove irrelevant validation of challenge input spending conditions
Checking the challenge input is able to spend a fabricated OP_TRUE input
doesn't verify anything useful. The challenge input is intentionally
unspendable to begin with.

The witness_utxo in the psbt for the challenge input is there for
signers, not validators.

The check that the challenge input txid is properly constructed is the
important one.