Hard code signature verification test for simplicity and determinism
Signature hex was obtained by running the previous version of the test,
but with `println!("signature: {signature}"); panic!();` inserted before
the `Witness::new()` line. The output was collected by running the test
as follows `cargo test invalid_signature -- --show-output`.

ECDSA signature generation appears to be deterministic on my machine so
you ought to be able to verify this new constant yourself.